pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jobmatch-de"
version = "0.1.0"
description = "AI-powered job matching CLI for the German market"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Adrian"}
]
keywords = ["job-search", "germany", "ai", "cli", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pdfplumber>=0.10.0",
    "google-search-results>=2.4.2",
    "google-genai>=1.0.0",
    "pydantic>=2.5.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
jobmatch-de = "jobmatch_de.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["jobmatch_de*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
